requirements-dev.txt
-r requirements.txt

# Linting & Testing
black==24.3.0
flake8==7.0.0
pylint==3.1.0
mypy==1.9.0
django-stubs==4.2.7
coverage==7.5.0

# Pre-commit
pre-commit==3.7.0 